Move type-c notification tests into own file
